requirements.txt
influxdb-client
numpy
pandas
scikit-learn
paho-mqtt
